# target/classes/ext/v1/swagger.json
{"swagger":"2.0","info":{"title":"EMBL-EBI REST API","version":"1","description":"Last updated date: **2021-12-19**","contact":{"name":"EMBL-EBI API Support","url":"httpss://embl.org","email":"support@embl.org"},"x-logo":{"url":"embl-logo.png","backgroundColor":"#ffffff"}},"schemes":["https"],"tags":[{"name":"Introduction","description":"The EMBL APIv1 is a REST API.\n\nBy utilizing the EMBL APIv1, you hereby agree to be bound by **EMBL-EBI** **[Terms of Service](www.embl.org/terms-of-service)**."},{"name":"General information","description":"## EMBL-EBI API Model Concept\nA common nomenclature is used to describe human and mouse genes. This provides the basis for communication and shared understanding in the scientific community when discussing genetic results and provides interoperability between data resources. The authority for human gene nomenclature is HGNC (https://www.genenames.org) located at EBI, while the authority for mouse gene nomenclature is the Mouse Genome Informatics resource (MGI) (http://www.informatics.jax.org). These authorities approve and assign a name, a unique symbol (often a short-form abbreviation of the name), and an identifier to each gene. Although the first complete sequences of the mouse and human genomes were obtained over 20 years ago genes are still being renamed and even discovered. The identifiers which are associated with the DNA sequence of the gene will stay the same even if the nomenclature is updated so are often preferred when databases exchange information, while biologists and clinicians tend to use the gene names and symbols. The nomenclature authorities publish updates on a weekly basis (MGI) and daily basis (HGNC) that provide both the current nomenclature for the genes and the previous symbols associated with the gene, which are referred to as gene synonyms.\n\nAt IMPC we generate a reference database every week that integrates the information published by MGI and HGNC to support our web services. The reference database contains information about the human and mouse genes and their synonyms, predictions about the relationship between human and mouse genes, as well as some additional information on diseases.\n\n"},{"name":"mouse gene","description":"The information of Mouse gene\n"},{"name":"mouse gene relations","description":"The information of Mouse gene and relation between mouse and human genes\n"}],"basePath":"/api/v1","consumes":["application/json"],"produces":["application/json"],"paths":{"/mouse_gene/{id}":{"parameters":[{"name":"mouse_gene_id","in":"path","description":"ID of mouse gene","required":true,"type":"number"}],"get":{"operationId":"get_mouse_gene_detail_by_id","summary":"GET:  Mouse Gene's detail by id","description":"Return the full Mouse Gene detail","tags":["mouse gene"],"responses":{"200":{"description":"Mouse Gene's detail","schema":{"type":"object","properties":{"id":{"type":"number"},"name":{"type":"string","description":"The name of Mouse gene","example":"RIKEN cDNA 1700088E04 gene"},"symbol":{"type":"string","description":"The symbol of Mouse gene","example":"1700088E04Rik"},"mgi_gene_acc_id":{"type":"string","description":null,"example":"MGI:1920774"}}}}}}},"/mouse_gene":{"parameters":[{"name":"symbol","in":"query","description":"The gene's symbol. e.g. Fgf8","required":false,"type":"string"},{"name":"mgi_gene_acc_id","in":"query","description":"The gene's mgi_gene_acc_id. e.g. MGI:99604","required":false,"type":"string"},{"name":"synonym","in":"query","description":"The gene's synonym. e.g. Aigf","required":false,"type":"string"}],"get":{"operationId":"get_mouse_gene_detail","summary":"GET:  Mouse Gene's detail by query","description":"Return the full Mouse Gene detail by query param `symbol` or `mgi_gene_acc_id` or `synonym`. Just used one param at a same time","tags":["mouse gene"],"responses":{"200":{"description":"Mouse Gene's detail with mouse gene symbols and synonyms","schema":{"type":"object","properties":{"id":{"type":"number"},"name":{"type":"string","description":"The name of Mouse gene","example":"RIKEN cDNA 1700088E04 gene"},"symbol":{"type":"string","description":"The symbol of Mouse gene","example":"1700088E04Rik"},"mgi_gene_acc_id":{"type":"string","description":null,"example":"MGI:1920774"},"synonyms":{"type":"array","description":"The list of Mouse gene synonyms","items":{"type":"string","example":"Aigf"}}}}}}}},"/mouse_gene/relations":{"parameters":[{"name":"symbol","in":"query","description":"The gene's symbol. e.g. Fgf8","required":false,"type":"string"},{"name":"mgi_gene_acc_id","in":"query","description":"The gene's mgi_gene_acc_id. e.g. MGI:99604","required":false,"type":"string"}],"get":{"operationId":"get_mouse_gene_relations","summary":"GET: The Mouse Gene and Human Gene relation by query","description":"Return the Mouse Gene detail and relation between mouse and human genes by query param `symbol` or `mgi_gene_acc_id`. Just used one param at a same time","tags":["mouse gene relations"],"responses":{"200":{"description":"The Mouse Gene detail and relation between mouse and human genes","schema":{"type":"object","properties":{"id":{"type":"number"},"name":{"type":"string","description":"The name of Mouse gene","example":"RIKEN cDNA 1700088E04 gene"},"symbol":{"type":"string","description":"The symbol of Mouse gene","example":"1700088E04Rik"},"mgi_gene_acc_id":{"type":"string","description":null,"example":"MGI:1920774"},"human_gene_relation_details":{"type":"array","items":{"type":"object","properties":{"human_gene":{"type":"object","properties":{"human_gene_id":{"type":"number","description":"The id of Human gene","example":2311}}},"support_count":{"type":"number","description":"The number of algorithms and predictions supporting the relationship","example":20}}}}}}}}}}}}